Move private functions back
